+++ src/main/java/org/zipli/socknet/exception/chat/UpdateChatException.java
@@ -2,9 +2,9 @@ package org.zipli.socknet.exception.chat;
 
 public class UpdateChatException extends RuntimeException{
 
-    private final long numberException;
+    private final int numberException;
 
-    public UpdateChatException(String message, long numberException) {
+    public UpdateChatException(String message, int numberException) {
         super(message);
         this.numberException = numberException;
     }
